Renamed file properly, corrected unit test config
