+++ python/src/ACEtk.python.cpp
@@ -106,6 +106,7 @@ namespace block {
   void wrapPhotoatomicElectronCrossSectionBlock( python::module&, python::module& );
   void wrapPhotoatomicElectronExcitationBlock( python::module&, python::module& );
   void wrapPhotoatomicElectronBremsstrahlungBlock( python::module&, python::module& );
+  void wrapPhotoelectricCrossSectionBlock( python::module&, python::module& );
 }
 
 // declarations - ACE table types
@@ -238,6 +239,7 @@ PYBIND11_MODULE( ACEtk, module ) {
   block::wrapPhotoatomicElectronCrossSectionBlock( module, viewmodule );
   block::wrapPhotoatomicElectronExcitationBlock( module, viewmodule );
   block::wrapPhotoatomicElectronBremsstrahlungBlock( module, viewmodule );
+  block::wrapPhotoelectricCrossSectionBlock( module, viewmodule );
 
   // wrap ACE table types
   wrapContinuousEnergyTable( module, viewmodule );
